perf(icons): scope workflow icon registration to cut webkit CI memory
The stories and unit-test pages each imported the full elements.js barrel,
registering all 413 <swc-icon-*> elements in every icons page the vitest
(webkit) run loads. Register only the icons each light page renders; the
Gallery story lazy-loads the barrel via a loader (it is excluded from the
runners), and the full-set manifest coverage test moves to its own page.
